Organise DukeException and subclasses into exceptions package and updated affected classes accordingly
