+++ src/main/java/duke/Duke.java
@@ -1,4 +1,6 @@
 package duke;
+import duke.exceptions.DukeException;
+
 import java.util.Scanner;
 
 public class Duke {
