--- src/main/java/duke/DukeException.java
@@ -1,7 +1,0 @@
-package duke;
-
-abstract class DukeException extends Exception{
-    public DukeException(String msg) {
-        super(msg);
-    }
-}
