+++ src/main/java/duke/Ui.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.exceptions.DukeException;
+
 import java.util.ArrayList;
 
 public class Ui {
