+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public abstract class DukeException extends Exception{
+    public DukeException(String msg) {
+        super(msg);
+    }
+}
