+++ src/main/java/duke/exceptions/DukeListOOBException.java
@@ -1,6 +1,4 @@
-package duke;
-
-import duke.DukeException;
+package duke.exceptions;
 
 public class DukeListOOBException extends DukeException {
     private static final String MSG = "The list does not contain an entry of index %d.\n";
